Removing thresholds for clocks and util rates
The warning/critical thresholds for the two metrics are not required as
problems concerning the clock are captured by the throttle reasons. For
util rates it is the case that they can vary a lot. Both sensors are
kept in the performance data and can be used as historical data.